Only process messages for instances tracked in the database
Before the new spot model, we would have to look up the values using the
api if this was the first time the instance had been seen.  This is
because fields like the instance ID would only be known when the spot
request converted into an instance.  Now that we know the instance id
immediately after running the runInstances method on EC2, we have
everything we need to insert the instance in the database.  This means
we can look up in the DB to know whether a given message is needing to
be processed or not.  The cloudwatchlog table will now only contain
tracked instance events, and we'll no longer blindly delete any message
which arrives with a non-pending, non-running state.